[ref] Integrate tray icon in the app
