+++ cozyfuse/interface/CozyFrame.py
@@ -1,16 +1,52 @@
-#!/usr/bin/env python
 # -*- coding: UTF-8 -*-
 #
-# generated by wxGlade 0.6.8 on Thu Aug  7 12:26:35 2014
+# generated by wxGlade 0.6.8 on Thu Aug  7 12:49:06 2014
 #
 
 import wx
-
 # begin wxGlade: dependencies
-import gettext
 # end wxGlade
 
 # begin wxGlade: extracode
+TRAY_TOOLTIP = 'System Tray Demo'
+TRAY_ICON = 'icon/icon.png'
+SMALL_TRAY_ICON = 'icon/small_icon.png'
+
+
+def create_menu_item(menu, label, func):
+    item = wx.MenuItem(menu, -1, label)
+    menu.Bind(wx.EVT_MENU, func, id=item.GetId())
+    menu.AppendItem(item)
+    return item
+
+
+class TaskBarIcon(wx.TaskBarIcon):
+    def __init__(self):
+        super(TaskBarIcon, self).__init__()
+        self.set_icon(TRAY_ICON)
+        self.Bind(wx.EVT_TASKBAR_LEFT_DOWN, self.on_left_down)
+
+    def CreatePopupMenu(self):
+        menu = wx.Menu()
+        create_menu_item(menu, 'Say Hello', self.on_hello)
+        menu.AppendSeparator()
+        create_menu_item(menu, 'Exit', self.on_exit)
+        return menu
+
+    def set_icon(self, path):
+        icon = wx.IconFromBitmap(wx.Bitmap(path))
+        self.SetIcon(icon, TRAY_TOOLTIP)
+
+    def on_left_down(self, event):
+        print 'Tray icon was left-clicked.'
+
+    def on_hello(self, event):
+        import debug
+        self.set_icon(SMALL_TRAY_ICON)
+        print 'Hello, world!'
+
+    def on_exit(self, event):
+        wx.CallAfter(self.Destroy)
 # end wxGlade
 
 
@@ -50,7 +86,7 @@ class CozyFrame(wx.Dialog):
         # begin wxGlade: CozyFrame.__set_properties
         self.SetTitle(_("Cozy client configuration"))
         _icon = wx.EmptyIcon()
-        _icon.CopyFromBitmap(wx.Bitmap("/home/kload/dev/cozy-fuse-linux/cozyfuse/interface/icon/icon.png", wx.BITMAP_TYPE_ANY))
+        _icon.CopyFromBitmap(wx.Bitmap("icon/icon.png", wx.BITMAP_TYPE_ANY))
         self.SetIcon(_icon)
         self.SetSize((400, 500))
         self.label_window_title.SetFont(wx.Font(11, wx.DEFAULT, wx.NORMAL, wx.BOLD, 0, ""))
@@ -120,12 +156,3 @@ class CozyFrame(wx.Dialog):
         event.Skip()
 
 # end of class CozyFrame
-if __name__ == "__main__":
-    gettext.install("app") # replace with the appropriate catalog name
-
-    app = wx.PySimpleApp(0)
-    wx.InitAllImageHandlers()
-    cozy_frame = CozyFrame(None, wx.ID_ANY, "")
-    app.SetTopWindow(cozy_frame)
-    cozy_frame.Show()
-    app.MainLoop()
