+++ cozyfuse/interface/app.py
@@ -1,0 +1,22 @@
+#!/usr/bin/env python
+# -*- coding: UTF-8 -*-
+#
+# generated by wxGlade 0.6.8 on Thu Aug  7 12:49:06 2014
+#
+
+# This is an automatically generated file.
+# Manual changes will be overwritten without warning!
+
+import wx
+import gettext
+from CozyFrame import CozyFrame
+
+if __name__ == "__main__":
+    gettext.install("app") # replace with the appropriate catalog name
+
+    app = wx.PySimpleApp(0)
+    wx.InitAllImageHandlers()
+    cozy_frame = CozyFrame(None, wx.ID_ANY, "")
+    app.SetTopWindow(cozy_frame)
+    cozy_frame.Show()
+    app.MainLoop()
